This commit adds the function to prepare an SCPPrepare message in Node.py and adds this function as an event in the Simulator.
The changes affect:

Node.py: 'prepare_ballot_msg' is defined which retrieves a confirmed message from the node's nomination state and makes an SCPPrepare message from it. It then adds it to its state and broadcast flag for SCPPrepare phase. This is also logged.

FBAConsensus.py: Added the 'prepare_ballot_msg' event to the Simulator.

Simulator.py: Added the 'prepare_ballot_msg' event