This commits adds functionality to process SCPPrepare messages and abort ballots.
A new function 'process_prepare_ballot_message' is made to process a received SCPPrepare message. Depending on the case the messaged ballot will be added to voted, or to aborted. Another case is that the node will abort all ballots smaller and with a different Value.

The 'abort_ballots' function aborts all voted ballots which are smaller and have a different value than the input ballot given.

These functions are needed to process received SCPPrepare messages in the protocol.