This commits adds the prepare SCPCommit functionalitz
The retrieve_confirmed_prepare_ballot retrieves a confirmed SCPPrepare message from a peer from the node's QuorumSet.

The 'prepare_SCPCommit_msg' retrieves a confirmed SCPPrepare message, checks if it is not None, initalises an SCPCommit message, and adds to its broadcast flag for peers to retrieve in later functionality to be implemented.

The Simulator and FBAConsensus classes have been edited to add this event