This commits adds functionality SCPExternalize data structure, and the prepare event for this phase.
The 'prepare_Externalize_msg' retrieves a confirmed commit ballot, creates an SCPExternalize message and stores it in its state and broadcast flag for peers.

The functions to receive and prepare ballots in the Prepare phase have also been edited to check if ballots have been finalised, before they are prepared. This is to ensure the Externalize phase is being used by nodes to track finalised slots.